Add OnExec()
syscall.Exec() will unapologetically replace/kill the current process.
This is good in many cases, but in some cases you really want to run
some code before it does so.

In my case, the Go backend also starts a webpack nodejs process to
rebuild the frontend on changes, but the node process will keep running
as an orphan because the parent died. With OnExec() the parent can
instruct the children to kill themselves. There are probably some other
use cases as well.

It would be nicer to add this to reload.Do(), but we can't really do
that without an incompatible change.